% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create Summary Dashboard}
\usage{
create_dashboard(results, problem_description = "")
}
\arguments{
\item{results}{List of result objects}

\item{problem_description}{Description of the problem}
}
\value{
Combined grid of plots
}
\description{
Create comprehensive dashboard with all metrics
}
