% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_brent_animation}
\alias{plot_brent_animation}
\title{Plot Brent's Method - Hybrid Approach Animation}
\usage{
plot_brent_animation(result, f)
}
\arguments{
\item{result}{Result object from brent_method}

\item{f}{Function being minimized}
}
\value{
plotly animation object
}
\description{
Shows Brent's combination of methods
}
