% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot Algorithm Comparison}
\usage{
plot_comparison(results)
}
\arguments{
\item{results}{List of result objects}

\item{metric}{Which metric to plot: "function_value", "x_value", or "step_size"}
}
\value{
ggplot object
}
\description{
Compare multiple algorithms on same plot
}
