% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_fixedpoint_animation}
\alias{plot_fixedpoint_animation}
\title{Plot Fixed-Point Iteration - Cobweb Diagram Animation}
\usage{
plot_fixedpoint_animation(result, f)
}
\arguments{
\item{result}{Result object from fixed_point_iteration}

\item{f}{Function being solved (will create g(x) = x - f(x))}
}
\value{
plotly animation object
}
\description{
Shows cobweb plot with y=g(x) and y=x
}
