% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_golden_animation}
\alias{plot_golden_animation}
\title{Plot Golden Section Search - Interval Reduction Animation}
\usage{
plot_golden_animation(result, f)
}
\arguments{
\item{result}{Result object from golden_section_search}

\item{f}{Function being minimized}
}
\value{
plotly animation object
}
\description{
Shows golden ratio points and interval shrinking
}
