% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_newton_animation}
\alias{plot_newton_animation}
\title{Plot Newton-Raphson Method - Tangent Lines Animation}
\usage{
plot_newton_animation(result, f, df = NULL)
}
\arguments{
\item{result}{Result object from newton_raphson}

\item{f}{Function being solved}

\item{df}{Derivative function (will compute numerically if NULL)}
}
\value{
plotly animation object
}
\description{
Shows tangent lines at each iteration leading to next guess
}
