% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_parabolic_animation}
\alias{plot_parabolic_animation}
\title{Plot Parabolic Interpolation - Parabola Fitting Animation}
\usage{
plot_parabolic_animation(result, f)
}
\arguments{
\item{result}{Result object from parabolic_interpolation}

\item{f}{Function being minimized}
}
\value{
plotly animation object
}
\description{
Shows parabola fitted through points
}
