% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_secant_animation}
\alias{plot_secant_animation}
\title{Plot Secant Method - Secant Lines Animation}
\usage{
plot_secant_animation(result, f)
}
\arguments{
\item{result}{Result object from secant_method}

\item{f}{Function being solved}
}
\value{
plotly animation object
}
\description{
Shows secant lines connecting consecutive points
}
