% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solve_minimize}
\alias{solve_minimize}
\title{Solve Minimization Problem}
\usage{
solve_minimize(f, interval, methods = "all", tol = 1e-06, max_iter = 100)
}
\arguments{
\item{f}{Function to minimize}

\item{interval}{Search interval c(a, b)}

\item{methods}{Vector of methods to try (default: "all")}

\item{tol}{Tolerance (default: 1e-6)}

\item{max_iter}{Maximum iterations (default: 100)}
}
\value{
List of results from all methods
}
\description{
Run all optimization algorithms on a function
}
